+++ NEWS.md
@@ -1,10 +1,11 @@
 # tarchetypes 0.6.0.9000
 
+* Add new functions `tar_quarto()` and `tar_quarto_rep()` (and "raw" versions) for Quarto documents and projects in pipelines (#89).
 * Temporarily change `root.dir` when scanning for dependencies so `knitr` child documents work (#93, @mutlusun).
 * Use `format = "rds"` for `target_batch` in `tar_map_rep_raw()` so the global format option does not mess up the pipeline.
 * Handle non-atomic length-one list columns in `tar_append_static_values()`.
 * Allow `tar_render()` to work with just one row of parameters (#96, #97, @ugoebel73).
-* Remove deps and collect garbage 
+* Remove dependencies and collect garbage before running reports.
 
 # tarchetypes 0.6.0